BackendAgri/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MongoDB and ODM
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0

# Redis for caching
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-decouple==3.8

# Cloud Storage (choose one or multiple)
# Google Cloud Storage
google-cloud-storage==2.10.0
# Azure Blob Storage (uncomment if using Azure)
# azure-storage-blob==12.19.0
# AWS S3 (uncomment if using AWS)
# boto3==1.34.0

# Image processing for plant diagnosis
Pillow==10.1.0

# HTTP client for external API calls
httpx==0.25.2
aiohttp==3.9.1

# Background tasks (optional)
celery==5.3.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.0

# Logging and monitoring
structlog==23.2.0

# Date and time handling
python-dateutil==2.8.2

# AI/ML libraries (add as needed)
# openai==1.3.0
# langchain==0.0.300
# transformers==4.35.0